pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-trend-digest"
version = "0.1.0"
description = "Daily AI trend digest generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "feedparser>=6.0.11",
  "httpx>=0.27.0",
  "lxml>=5.2.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "trafilatura>=1.9.0",
  "readability-lxml>=0.8.1",
  "openai>=1.40.0",
  "google-genai>=0.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.5.0"]
local = ["llama-cpp-python[server]>=0.2.90"]

[tool.setuptools.packages.find]
include = ["digest*"]
